requirements.txt
ultralytics==8.1.33
torchreid
torch>=2.0.0
opencv-python
matplotlib
numpy
Pillow
scikit-learn
